Move get_thread_id to GeantUtils
