+++ app/celer-g4/RootIO.cc
@@ -20,7 +20,7 @@
 
 #include "corecel/Macros.hh"
 #include "corecel/io/Logger.hh"
-#include "celeritas/ext/GeantSetup.hh"
+#include "celeritas/ext/GeantUtils.hh"
 #include "accel/ExceptionConverter.hh"
 #include "accel/SetupOptions.hh"
