+++ app/celer-g4/TimerOutput.cc
@@ -10,7 +10,7 @@
 #include "corecel/Assert.hh"
 #include "corecel/Macros.hh"
 #include "corecel/io/JsonPimpl.hh"
-#include "accel/SetupOptions.hh"
+#include "celeritas/ext/GeantUtils.hh"
 
 #if CELERITAS_USE_JSON
 #    include <nlohmann/json.hpp>
@@ -63,7 +63,7 @@ void TimerOutput::output(JsonPimpl* j) const
  */
 void TimerOutput::RecordActionTime(MapStrReal&& time)
 {
-    size_type thread_id = GetThreadID();
+    size_type thread_id = get_thread_id();
     CELER_ASSERT(thread_id < action_time_.size());
     action_time_[thread_id] = std::move(time);
 }
@@ -74,7 +74,7 @@ void TimerOutput::RecordActionTime(MapStrReal&& time)
  */
 void TimerOutput::RecordEventTime(real_type time)
 {
-    size_type thread_id = GetThreadID();
+    size_type thread_id = get_thread_id();
     CELER_ASSERT(thread_id < event_time_.size());
     event_time_[thread_id].push_back(time);
 }
