+++ src/celeritas/ext/GeantSetup.hh
@@ -80,9 +80,6 @@ class GeantSetup
 
 //---------------------------------------------------------------------------//
 // FREE FUNCTIONS
-//---------------------------------------------------------------------------//
-// Get the number of threads in a version-portable way
-int get_num_threads(G4RunManager const&);
 
 //---------------------------------------------------------------------------//
 // INLINE DEFINITIONS
@@ -107,11 +104,6 @@ inline GeantSetup::GeantSetup(std::string const&, Options)
 inline GeantSetup::~GeantSetup() = default;
 
 inline void GeantSetup::RMDeleter::operator()(G4RunManager*) const
-{
-    CELER_ASSERT_UNREACHABLE();
-}
-
-inline int get_num_threads(G4RunManager const&)
 {
     CELER_ASSERT_UNREACHABLE();
 }
